refactor: move components to appropriate folders
