+++ front/src/App.jsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 import Style from './App.module.scss';
-import Header from '../Header/Header';
-import AddRecipe from '../AddRecipe/AddRecipe';
-import Footer from '../Footer/Footer';
+import Header from './components/Layout/Header/Header';
+import AddRecipe from './components/Form/AddRecipe/AddRecipe';
+import Footer from './components/Layout/Footer/Footer';
 // import Toast from '../Toast/Toast';
-import Modal from '../Modal/Modal';
+import Modal from './components/Modal/Modal';
 
 const App = () => (
   <div className={Style.app}>
