+++ front/src/components/Form/Legal/Legal.jsx
