+++ front/src/components/Form/Legal/Legal.module.scss
