+++ front/src/components/Form/MainInfo/mainInfo.module.scss
@@ -114,7 +114,7 @@
 }
 
 .star {
-  background-image: url('../../images/star.svg');
+  background-image: url('../../../images/star.svg');
   background-position: 50%;
   background-size: contain;
   background-repeat: no-repeat;
@@ -150,28 +150,28 @@
 }
 
 .buttonPlus {
-  background-image: url('../../images/plusButton.svg');
+  background-image: url('../../../images/plusButton.svg');
   background-position: 50%;
   background-size: contain;
   background-repeat: no-repeat;
 
   &:hover {
-    background-image: url('../../images/plusButton-active.svg');
+    background-image: url('../../../images/plusButton-active.svg');
   }
 }
 
 .buttonMinus {
-  background-image: url('../../images/minusButton.png');
+  background-image: url('../../../images/minusButton.png');
   background-position: 50%;
   background-size: contain;
   background-repeat: no-repeat;
 
   &:hover {
-    background-image: url('../../images/minusButton-active.svg');
+    background-image: url('../../../images/minusButton-active.svg');
   }
 
   &:disabled {
-    background-image: url('../../images/minusButton-disabled.png');
+    background-image: url('../../../images/minusButton-disabled.png');
     cursor: auto;
   }
 }
@@ -318,7 +318,7 @@
   margin: 0;
   padding: 0;
   margin-left: 10px;
-  list-style-image: url('../../images/list-marker.svg');
+  list-style-image: url('../../../images/list-marker.svg');
 }
 
 .reqlist_item {
