+++ front/src/components/Form/ProgressBar/ProgressBar.jsx
@@ -2,7 +2,7 @@
 /* eslint-disable jsx-a11y/anchor-is-valid */
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { progressBarSteps } from '../../utils/constants';
+import { progressBarSteps } from '../../../utils/constants';
 
 import styles from './ProgressBar.module.scss';
