+++ front/src/components/Form/Tooltip/TooltipDifficultyContent/TooltipDifficultyContent.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Style from './TooltipDifficultyContent.module.scss';
-import starFilled from '../../../images/star-filled.svg';
+import starFilled from '../../../../images/star-filled.svg';
 
 const TooltipDifficultyContent = () => (
   <>
