+++ front/src/components/Form/Tooltip/TooltipDifficultyContent/TooltipDifficultyContent.module.scss
