+++ front/src/components/Layout/Header/Header.module.scss
@@ -31,7 +31,7 @@
 
   .buttonMenu {
     display: inline-block;
-    background: url('../../images/icon-burger-menu.svg') center;
+    background: url('../../../images/icon-burger-menu.svg') center;
     width: 32px;
     height: 32px;
     border: none;
