+++ front/src/components/Layout/Logo/Logo.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Style from './Logo.module.scss';
-import logo from '../../images/logo.svg';
+import logo from '../../../images/logo.svg';
 
 const Logo = () => (
   <img className={Style.logo} src={logo} alt="Логотип приложения" />
