+++ front/src/components/UI/CloseButton/CloseButton.module.scss
@@ -2,10 +2,10 @@
   width: 20px;
   height: 20px;
   background-color: transparent;
-  background-image: url('../../images/close_btn.svg');
+  background-image: url('../../../images/close_btn.svg');
   border: none;
-  opacity: .6;
-  transition: opacity .6s;
+  opacity: 0.6;
+  transition: opacity 0.6s;
   cursor: pointer;
 
   &:hover {
