+++ front/src/components/UI/PhotoButton/PhotoButton.module.scss
