+++ front/src/pages/FormPage.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const FormPage = () => <div>Здесь будет собрана страница с формой</div>;
+
+export default FormPage;
