+++ front/src/pages/RecipeListPage.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const RecipeListPage = () => (
+  <div>Здесь будет собрана страница с рецептами</div>
+);
+
+export default RecipeListPage;
